+++ core/src/main/java/ai/timefold/solver/core/api/domain/lookup/LookUpStrategyType.java
@@ -8,7 +8,10 @@ import ai.timefold.solver.core.api.score.director.ScoreDirector;
  * Determines how {@link ScoreDirector#lookUpWorkingObject(Object)} maps
  * a {@link ProblemFactCollectionProperty problem fact} or a {@link PlanningEntity planning entity}
  * from an external copy to the internal one.
+ * 
+ * @deprecated When multi-threaded solving, ensure your domain classes use @{@link PlanningId} instead.
  */
+@Deprecated(forRemoval = true, since = "1.10.0")
 public enum LookUpStrategyType {
     /**
      * Map by the same {@link PlanningId} field or method.
