+++ core/src/main/java/ai/timefold/solver/core/impl/domain/lookup/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * @deprecated When multi-threaded solving,
+ *             ensure your domain classes use @{@link ai.timefold.solver.core.api.domain.lookup.PlanningId} instead.
+ */
+@Deprecated(forRemoval = true, since = "1.10.0")
+package ai.timefold.solver.core.impl.domain.lookup;
